// client/src/store/model/socketUpdate.ts
/*
 * Copyright (C) 2020 Evgenia Lazareva
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

 import { Contact } from "./contact";

export enum SocketEvents {
  StatusUpdated = 'statusUpdated',
  StatusDeleted = 'statusDeleted',

  ContactUpdate = 'contactUpdate',
  ContactDelete = 'contactDelete'
}

export enum ContactEventTypes {
  ContactRequest = 'contactRequest',
  ContactRequestApproved = 'contactRequestApproved',
  ContactRequestRejected = 'contactRequestRejected',
  ContactRequestWithdrawn = 'contactRequestWithdrawn',
  ContactDeleted = 'contactDeleted',
  ContactNameUpdated = 'contactNameUpdated',
  ContactTrustUpdated = 'contactTrustUpdated',
  ContactDoneActionsUpdated = 'contactDoneActionsUpdated'
}

export interface SocketContactUpdate {
  id: string;
  contact: Contact;
  updateType: ContactEventTypes;  
}